requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
aws-sam-translator==1.92.0
blinker==1.4
boto3==1.35.57
botocore==1.35.57
cfn-lint==1.18.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.7
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
exceptiongroup==1.2.2
gyp==0.1
httplib2==0.20.2
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.0.0
jeepney==0.7.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
oauthlib==3.2.0
packaging==24.2
patool==2.4.0
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.2.0
pytest==8.3.3
pytest-django==4.9.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==5.4.1
referencing==0.35.1
regex==2024.11.6
rpds-py==0.21.0
s3transfer==0.10.3
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
systemd-python==234
tomli==2.0.2
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.3
wadllib==1.3.6
zipp==1.0.0
